refactor: fold indexer.ts into utils.ts
Abstract constants and helpers belong in utils.ts, not a bespoke module.
Moves the Indexer type, INDEXER_HOSTS map, selectedIndexer, and
electrumBridge into utils.ts (alongside the other bridge helpers) and
updates the three importers; no behavior change.
